+++ app/src/main/java/com/example/tracker/ui/SavedWalkViewHolder.kt
@@ -6,7 +6,7 @@ import kotlinx.android.synthetic.main.address_and_distance_item.view.*
 
 class SavedWalkViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
-    fun bind(location: String, distance: Int) {
+    fun bind(location: String, distance: Double) {
         itemView.inItemLocation.text = location
         itemView.inItemDistance.text = distance.toString()
     }
